unit_test_gen/requirements.txt
openai>=1.99.9
python-dotenv
faiss-cpu
sentence-transformers
markdown
pymilvus==2.3.5



